+++ packages/spacecat-shared-data-access/src/v2/models/site-candidate/site-candidate.schema.js
@@ -54,6 +54,6 @@ const schema = new SchemaBuilder(SiteCandidate, SiteCandidateCollection)
   .addAttribute('updatedBy', {
     type: 'string',
   })
-  .addAllIndexWithComposite('baseURL');
+  .addAllIndex(['baseURL']);
 
 export default schema.build();